# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Storefront catalog, store locations, and owner-managed site content
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  priceInr: number | null;
  imageUrls: string[];
  sizes: string[];
  colors: string[];
  /** @nullable */
  category: string | null;
  /** @nullable */
  collection: string | null;
  /** @nullable */
  stock: number | null;
  isPublished: boolean;
  isFeatured: boolean;
  isNewArrival: boolean;
}

export interface ProductInput {
  /** @minLength 1 */
  name: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  priceInr?: number | null;
  imageUrls?: string[];
  sizes?: string[];
  colors?: string[];
  /** @nullable */
  category?: string | null;
  /** @nullable */
  collection?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  stock?: number | null;
  isPublished?: boolean;
  isFeatured?: boolean;
  isNewArrival?: boolean;
}

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  priceInr?: number | null;
  imageUrls?: string[];
  sizes?: string[];
  colors?: string[];
  /** @nullable */
  category?: string | null;
  /** @nullable */
  collection?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  stock?: number | null;
  isPublished?: boolean;
  isFeatured?: boolean;
  isNewArrival?: boolean;
}

export interface Store {
  id: number;
  name: string;
  /** @nullable */
  address: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  whatsapp: string | null;
  /** @nullable */
  openingHours: string | null;
  /** @nullable */
  mapsUrl: string | null;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  description: string | null;
}

export interface StoreInput {
  /** @minLength 1 */
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  openingHours?: string | null;
  /** @nullable */
  mapsUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface StoreUpdate {
  /** @minLength 1 */
  name?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  openingHours?: string | null;
  /** @nullable */
  mapsUrl?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  description?: string | null;
}

export interface SiteSettings {
  brandName: string;
  heroHeadline: string;
  /** @nullable */
  heroSubheading: string | null;
  /** @nullable */
  heroImageUrl: string | null;
  instagramUrl: string;
  categories: string[];
  /** @nullable */
  featuredCollectionTitle: string | null;
  /** @nullable */
  featuredCollectionDescription: string | null;
  /** @nullable */
  aboutStory: string | null;
}

export interface SiteSettingsInput {
  /** @minLength 1 */
  brandName: string;
  /** @minLength 1 */
  heroHeadline: string;
  /** @nullable */
  heroSubheading: string | null;
  /** @nullable */
  heroImageUrl: string | null;
  instagramUrl: string;
  categories: string[];
  /** @nullable */
  featuredCollectionTitle: string | null;
  /** @nullable */
  featuredCollectionDescription: string | null;
  /** @nullable */
  aboutStory: string | null;
}

export interface AdminAccessStatus {
  signedIn: boolean;
  isAdmin: boolean;
}

export type ListProductsParams = {
query?: string;
category?: string;
collection?: string;
sort?: ListProductsSort;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  featured: 'featured',
  newest: 'newest',
  'price-asc': 'price-asc',
  'price-desc': 'price-desc',
} as const;