# lib/api-zod/src/generated/types/adminAccessStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Storefront catalog, store locations, and owner-managed site content
 * OpenAPI spec version: 0.1.0
 */

export interface AdminAccessStatus {
  signedIn: boolean;
  isAdmin: boolean;
}