# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Storefront catalog, store locations, and owner-managed site content
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  slug: string;
  name: string;
  /** @nullable */
  description: string | null;
  /** @nullable */
  priceInr: number | null;
  imageUrls: string[];
  sizes: string[];
  colors: string[];
  /** @nullable */
  category: string | null;
  /** @nullable */
  collection: string | null;
  /** @nullable */
  stock: number | null;
  isPublished: boolean;
  isFeatured: boolean;
  isNewArrival: boolean;
}