# lib/api-zod/src/generated/types/productUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Storefront catalog, store locations, and owner-managed site content
 * OpenAPI spec version: 0.1.0
 */

export interface ProductUpdate {
  /** @minLength 1 */
  name?: string;
  slug?: string;
  /** @nullable */
  description?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  priceInr?: number | null;
  imageUrls?: string[];
  sizes?: string[];
  colors?: string[];
  /** @nullable */
  category?: string | null;
  /** @nullable */
  collection?: string | null;
  /**
     * @minimum 0
     * @nullable
     */
  stock?: number | null;
  isPublished?: boolean;
  isFeatured?: boolean;
  isNewArrival?: boolean;
}