# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Storefront catalog, store locations, and owner-managed site content
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettings {
  brandName: string;
  heroHeadline: string;
  /** @nullable */
  heroSubheading: string | null;
  /** @nullable */
  heroImageUrl: string | null;
  instagramUrl: string;
  categories: string[];
  /** @nullable */
  featuredCollectionTitle: string | null;
  /** @nullable */
  featuredCollectionDescription: string | null;
  /** @nullable */
  aboutStory: string | null;
}