# lib/api-zod/src/generated/types/siteSettingsInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Storefront catalog, store locations, and owner-managed site content
 * OpenAPI spec version: 0.1.0
 */

export interface SiteSettingsInput {
  /** @minLength 1 */
  brandName: string;
  /** @minLength 1 */
  heroHeadline: string;
  /** @nullable */
  heroSubheading: string | null;
  /** @nullable */
  heroImageUrl: string | null;
  instagramUrl: string;
  categories: string[];
  /** @nullable */
  featuredCollectionTitle: string | null;
  /** @nullable */
  featuredCollectionDescription: string | null;
  /** @nullable */
  aboutStory: string | null;
}