# lib/api-zod/src/generated/types/store.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Storefront catalog, store locations, and owner-managed site content
 * OpenAPI spec version: 0.1.0
 */

export interface Store {
  id: number;
  name: string;
  /** @nullable */
  address: string | null;
  /** @nullable */
  phone: string | null;
  /** @nullable */
  whatsapp: string | null;
  /** @nullable */
  openingHours: string | null;
  /** @nullable */
  mapsUrl: string | null;
  /** @nullable */
  imageUrl: string | null;
  /** @nullable */
  description: string | null;
}